let n type from u16 to u32

+++ include/k_string.h
@@ -56,7 +56,7 @@ static inline int s_memcmp(const char *s1, const char *s2, int n)
   return 0;
 }
 
-void p_asm_memcpy(void *dest, void *src, u16 n);
-void p_asm_memset(void *dest, int c, u16 n);
+void p_asm_memcpy(void *dest, void *src, u32 n);
+void p_asm_memset(void *dest, int c, u32 n);
 
 #endif
